+++ lily/midi-item.cc
@@ -87,7 +87,7 @@ Midi_instrument::to_string () const
   SCM program = scm_call_1 (proc, ly_symbol2scm (audio_->str_.c_str ()));
   found = (program != SCM_BOOL_F);
   if (found)
-    program_byte = scm_to_int (program);
+    program_byte = (Byte) scm_to_int (program);
   else
     warning (_f ("no such MIDI instrument: `%s'", audio_->str_.c_str ()));
